requirements.txt
appnope==0.1.2
APScheduler==3.7.0
argon2-cffi==20.1.0
asgiref==3.3.1
async-generator==1.10
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
backcall==0.2.0
backtrader==1.9.76.123
bleach==3.3.0
Brotli==1.0.9
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==3.4.7
cycler==0.10.0
datanews==0.0.7
dateparser==1.0.0
decorator==5.0.6
defusedxml==0.6.0
Django==3.1.4
django-allauth==0.44.0
django-apscheduler==0.5.2
django-cors-headers==3.5.0
django-cors-middleware==1.5.0
django-jsonfield==1.4.1
django-rest-auth==0.9.5
django-token==0.1.3
djangorestframework==3.12.2
empyrical==0.5.5
entrypoints==0.3
Flask==1.1.2
gunicorn==20.0.4
hyperlink==21.0.0
idna==2.7
importlib-metadata==3.10.0
incremental==21.3.0
inflection==0.5.1
install==1.3.4
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==0.24
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==0.23
matplotlib==3.4.1
mistune==0.8.4
more-itertools==8.6.0
multitasking==0.0.9
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.19.4
oauthlib==3.1.0
packaging==20.9
pandas==1.1.4
pandas-datareader==0.9.0
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
plaid-python==5.0.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyfolio==0.9.2
Pygments==2.8.1
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyotp==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-binance==0.7.11
python-dateutil==2.8.1
python-dotenv==0.14.0
python3-openid==3.2.0
pytz==2020.4
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
Quandl==3.5.3
regex==2021.4.4
requests==2.20.0
requests-oauthlib==1.3.0
robin-stocks==2.0.3
scikit-learn==0.24.1
scipy==1.6.2
seaborn==0.11.1
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
TA-Lib==0.4.19
td-ameritrade-python-api==0.3.4
tda-api==0.7.1
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==4.0.2
urllib3==1.24.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
websockets==8.1
Werkzeug==1.0.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
yfinance==0.1.59
zipp==3.4.1
zope.interface==5.4.0
